Add TODO
Add a TODO comment so I won't forget to make the "getCurrentPosition" method return rotations instead of encoder ticks. The encoder to rotations logic SHOULD be included in each motor's API, so that we don't have to deal with converting motors with difference amounts of ticks per rotation to their rotations in our main autonomous or teleop programs.