Major API structure changes
Change the way our API structure works by making the ChassisAPI and interface between the main program and all other APIs. This commit adds methods to the ChassisAPI that allows it to run the basic functions of the other APIs. This keeps us from having to instantiate many classes when we're trying to create a program in a hurry. In addition, all of the values that are hard coded (like encoder ticks and wheel diameters) is now in the constants class. This allows us to find the variables we need to adjust much quicker and allows us to hard code them in once ins tead of in many different places.